# 03-Azure/01-04-AI/01_Agentic_SDLC/src/api-ts/api-swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Express API with Swagger",
        "version": "1.0.0",
        "description": "REST API documentation using Swagger/OpenAPI"
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Development server"
        }
    ],
    "paths": {
...
        "/api/carts/{cartId}": {
            "get": {
                "summary": "Get a cart aggregate",
                "tags": [
                    "Carts"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "cartId",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cart aggregate",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cart"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid cart ID"
                    },
                    "404": {
                        "description": "Cart not found"
                    }
                }
            }
        },
        "/api/carts/{cartId}/items": {
            "post": {
                "summary": "Add a product or increment its existing cart quantity",
                "tags": [
                    "Carts"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "cartId",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddCartItemRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated cart aggregate",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cart"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "404": {
                        "description": "Cart or product not found"
                    }
                }
            }
        },
        "/api/carts/{cartId}/items/{productId}": {
            "put": {
                "summary": "Replace a cart line quantity",
                "tags": [
                    "Carts"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "cartId",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "in": "path",
                        "name": "productId",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCartItemRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated cart aggregate",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cart"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "404": {
                        "description": "Cart line not found"
                    }
                }
            },
            "delete": {
                "summary": "Remove a product from a cart",
                "tags": [
                    "Carts"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "cartId",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "in": "path",
                        "name": "productId",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Cart line removed"
                    },
                    "400": {
                        "description": "Invalid path ID"
                    },
                    "404": {
                        "description": "Cart or cart line not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CartItem": {
                "type": "object",
                "required": [
                    "cartItemId",
                    "productId",
                    "name",
                    "imgName",
                    "quantity",
                    "unitPrice",
                    "lineTotal"
                ],
                "properties": {
                    "cartItemId": {
                        "type": "integer"
                    },
                    "productId": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "imgName": {
                        "type": "string",
                        "nullable": true
                    },
                    "quantity": {
                        "type": "integer",
                        "minimum": 1
                    },
                    "unitPrice": {
                        "type": "number",
                        "format": "double",
                        "minimum": 0
                    },
                    "lineTotal": {
                        "type": "number",
                        "format": "double",
                        "minimum": 0
                    }
                }
            },
            "Cart": {
                "type": "object",
                "required": [
                    "cartId",
                    "createdAt",
                    "items",
                    "itemCount",
                    "total"
                ],
                "properties": {
                    "cartId": {
                        "type": "integer"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CartItem"
                        }
                    },
                    "itemCount": {
                        "type": "integer",
                        "minimum": 0
                    },
                    "total": {
                        "type": "number",
                        "format": "double",
                        "minimum": 0
                    }
                }
            },
            "AddCartItemRequest": {
                "type": "object",
                "required": [
                    "productId",
                    "quantity"
                ],
                "additionalProperties": false,
                "properties": {
                    "productId": {
                        "type": "integer",
                        "minimum": 1
                    },
                    "quantity": {
                        "type": "integer",
                        "minimum": 1
                    }
                }
            },
            "UpdateCartItemRequest": {
                "type": "object",
                "required": [
                    "quantity"
                ],
                "additionalProperties": false,
                "properties": {
                    "quantity": {
                        "type": "integer",
                        "minimum": 1
                    }
                }
            },
...
        }
    },
...
}